#19 - Change stored query handling
- move stored queries from v1 rest controller to v2
